Made BaseAuto class (moved apriltag methods here) and changed BaseOpMode to use SampleMecanumDrive class for drive motors
