+++ Team417_PowerPlay/src/main/java/org/firstinspires/ftc/team417_PowerPlay/BaseOpMode.java
@@ -10,20 +10,15 @@ import org.firstinspires.ftc.team417_PowerPlay.drive.SampleMecanumDrive;
 
 abstract public class BaseOpMode extends LinearOpMode {
 
-    DcMotor motorFL;
-    DcMotor motorFR;
-    DcMotor motorBL;
-    DcMotor motorBR;
-    DcMotor leftEncoder;
-    DcMotor rightEncoder;
-    DcMotor frontEncoder;
     DcMotor motorArm;
 
     Servo grabberServo;
     Toggler grabberToggle;
-    ElapsedTime time;
+
+    SampleMecanumDrive drive;
     public static final double GRABBER_OPEN = 0.9;
     public static final double GRABBER_CLOSED = 0.4;
+    public static final double GRABBER_HALF_CLOSED = 0.8;
 
     public static final int MAX_ARM_POSITION = -1500;
     public static final int GRD_JUNCT_ARM_POSITION = -400;
@@ -32,57 +27,20 @@ abstract public class BaseOpMode extends LinearOpMode {
     public static final int MIN_ARM_POSITION = 0;
 
     public void initializeHardware() {
-        time = new ElapsedTime();
-        time.reset();
-        motorFL = hardwareMap.dcMotor.get("motorFL");
-        motorFR = hardwareMap.dcMotor.get("motorFR");
-        motorBL = hardwareMap.dcMotor.get("motorBL");
-        motorBR = hardwareMap.dcMotor.get("motorBR");
+        drive = new SampleMecanumDrive(hardwareMap);
+
         motorArm = hardwareMap.dcMotor.get("motorArm");
 
-        leftEncoder = hardwareMap.dcMotor.get("leftEncoder");
-        rightEncoder = hardwareMap.dcMotor.get("rightEncoder");
-        frontEncoder = hardwareMap.dcMotor.get("frontEncoder");
-
-        motorFL.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-        motorFR.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-        motorBL.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-        motorBR.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
         motorArm.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
 
-        leftEncoder.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-        rightEncoder.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-        frontEncoder.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-
-        motorFL.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-        motorFR.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-        motorBL.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-        motorBR.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
         motorArm.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
 
-        leftEncoder.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-        rightEncoder.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-        frontEncoder.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-
-        motorFR.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
-        motorFL.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
-        motorBL.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
-        motorBR.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
         motorArm.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
 
-        motorFL.setDirection(DcMotor.Direction.REVERSE);
-        motorBL.setDirection(DcMotor.Direction.REVERSE);
-        motorBR.setDirection(DcMotor.Direction.FORWARD);
-        motorFR.setDirection(DcMotor.Direction.FORWARD);
-
         grabberServo = hardwareMap.servo.get("grabberServo");
-        grabberServo.setPosition(GRABBER_CLOSED);
         grabberToggle = new Toggler();
 
-        motorFL.setPower(0);
-        motorFR.setPower(0);
-        motorBL.setPower(0);
-        motorBR.setPower(0);
+        drive.setMotorPowers(0, 0, 0, 0);
         motorArm.setPower(0);
     }
 
@@ -97,9 +55,6 @@ abstract public class BaseOpMode extends LinearOpMode {
         powerBL = y - x + turning;
         powerBR = y + x - turning;
 
-        motorFL.setPower(powerFL);
-        motorFR.setPower(powerFR);
-        motorBL.setPower(powerBL);
-        motorBR.setPower(powerBR);
+        drive.setMotorPowers(powerFL, powerBL, powerBR, powerFR);
     }
 }
