+++ Team417_PowerPlay/src/main/java/org/firstinspires/ftc/team417_PowerPlay/TeleOpCompetition.java
@@ -3,11 +3,11 @@ package org.firstinspires.ftc.team417_PowerPlay;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
 @TeleOp (name="Teleop Competition")
-public class TeleOpCompetition extends BaseTeleOp{
+public class TeleOpCompetition extends BaseTeleOp {
 
     @Override
     public void runOpMode() throws InterruptedException {
-        initializeHardware();
+        initializeTeleOp();
         telemetry.addLine("Ready for start.");
         telemetry.update();
         waitForStart();
